404 error pages (#1312)
* tag and wiki error page

* tag and wiki error page

* creted notfound page

* creted notfound page
